# docs/swagger.yaml
basePath: /
definitions:
  ErrorResponse:
    properties:
      message:
        type: string
      param:
        type: string
      value:
        type: string
    type: object
...
  models.Pitch:
    properties:
      cost:
        type: number
      maxPlayers:
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      pitchId:
        type: integer
      venueId:
        type: integer
    required:
    - cost
    - maxPlayers
    - name
    - venueId
    type: object
...
  models.Venue:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNo:
        type: string
      postcode:
        type: string
    type: object
...
  payloads.PitchRequest:
    properties:
      cost:
        type: number
      maxPlayers:
        maximum: 22
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - cost
    - maxPlayers
    - name
    type: object
...
  payloads.VenueAdminRequest:
    properties:
      userId:
        type: integer
      venueId:
        type: integer
    type: object
  payloads.VenueAdminResponse:
    properties:
      email:
        type: string
      forename:
        type: string
      surname:
        type: string
      venueAdminId:
        type: integer
    type: object
  payloads.VenueRequest:
    properties:
      address:
        maxLength: 100
        minLength: 3
        type: string
      city:
        maxLength: 50
        minLength: 3
        type: string
      email:
        maxLength: 100
        minLength: 10
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      phoneNo:
        maxLength: 15
        minLength: 11
        type: string
      postcode:
        maxLength: 8
        minLength: 5
        type: string
    required:
    - address
    - city
    - email
    - name
    - phoneNo
    - postcode
    type: object
  views.Venue:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNo:
        type: string
      pitches:
        items:
          $ref: '#/definitions/views.VenuePitchSummary'
        type: array
      postcode:
        type: string
    type: object
  views.VenuePitchSummary:
    properties:
      href:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: 127.0.0.1:3000
info:
  contact:
    email: nathan.mcalinden@unosquare.com
    name: Footpal Support
  description: This is swagger for Footpal
  termsOfService: http://swagger.io/terms/
  title: Footpal API
  version: "1.0"
paths:
...
  /venues:
    get:
      description: Retrieve all venues
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/views.Venue'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - venue
    post:
      description: Create new football venue
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.VenueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - venue
    put:
      description: Edit venue details
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.VenueRequest'
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Venue'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - venue
  /venues/{venueId}:
    delete:
      description: Remove venue from Footpal
      parameters:
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - venue
    get:
      description: Retrieve venue by venueId
      parameters:
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.Venue'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - venue
  /venues/{venueId}/admins:
    delete:
      description: Remove administrator from venue
      parameters:
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - venue
    get:
      description: Retrieve venue administrators
      parameters:
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/payloads.VenueAdminResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - venue
    post:
      description: Add new administrator to venue
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.VenueAdminRequest'
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/payloads.VenueAdminResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - venue
  /venues/{venueId}/pitches:
    get:
      description: Get Pitches by Venue
      parameters:
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Pitch'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - venue
    post:
      description: Add new pitch to existing venue
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.PitchRequest'
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Pitch'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - venue
  /venues/{venueId}/pitches/{pitchId}:
    delete:
      description: Remove pitch from venue
      parameters:
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      - description: Pitch ID
        in: path
        name: pitchId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - venue
    get:
      description: Get pitch info by venue
      parameters:
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pitch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - venue
    put:
      description: Edit pitch details
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.PitchRequest'
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      - description: Pitch ID
        in: path
        name: pitchId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pitch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - venue
  /venues/{venueId}/pitches/{pitchId}/timeslots:
    get:
      description: Retrieve all time slots by Pitch
      parameters:
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      - description: Pitch ID
        in: path
        name: pitchId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - venue
  /venues/{venueId}/timeslots:
    get:
      description: Retrieve all time slots by Venue
      parameters:
      - description: Venue ID
        in: path
        name: venueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - venue
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"